Bug/path injection attack fix

+++ src/backend/app/services/convert_service.py
@@ -5,6 +5,7 @@ from pathlib import Path
 import subprocess
 from fastapi import UploadFile, HTTPException
 from app.core.constants import CONVERTED_FILE_NAME
+import aiofiles
 
 logger = logging.getLogger(__name__)
 
@@ -13,18 +14,20 @@ async def convert_file_service(file: UploadFile):
     response_obj = {"extracted_text": ""}
 
     try:
-        filename: Path = Path(file.filename)
+        filename: Path = Path(file.filename) # type: ignore
         file_type: str = filename.suffix
         content: bytes = await file.read()
         extracted_text = ""
         page_count = 0
 
-        with open(f"{file.filename}", "wb") as f:  # saves to the build folder
-            f.write(content)
+        # Prevent path traversal by using only the file name, not the full path
+        safe_filename = os.path.basename(file.filename) # type: ignore
+        async with aiofiles.open(safe_filename, "wb") as f:
+            await f.write(content)
 
         # not always 100% successful if pdfs are complicated or encoded in a way that it cannot parse
         if file_type == ".pdf":
-            reader = pypdf.PdfReader(f"{file.filename}")
+            reader = pypdf.PdfReader(f"{safe_filename}")  # type: ignore
             page_count += len(reader.pages)
 
             for page in reader.pages:
@@ -40,26 +43,28 @@ async def convert_file_service(file: UploadFile):
             # Run Pandoc to convert the file from commandline
             command = [
                 "pandoc",
-                file.filename,
+                safe_filename,
                 "-o",
                 CONVERTED_FILE_NAME,
             ]
             completed_process = subprocess.run(command, check=True, capture_output=True)
 
             # Read the converted file
-            with open(CONVERTED_FILE_NAME, "rb") as f:
-                converted_data = f.read()
+            async with aiofiles.open(CONVERTED_FILE_NAME, "rb") as f:
+                converted_data = await f.read()
                 response_obj["extracted_text"] = converted_data
                 response_obj["pages"] = None
                 response_obj["file_type"] = "docx"
 
         try:
             # finally remove file:
-            os.remove(CONVERTED_FILE_NAME)
-            os.remove(f"{file.filename}")
+
+            if os.path.exists(CONVERTED_FILE_NAME):
+                os.remove(CONVERTED_FILE_NAME)
+            os.remove(f"{safe_filename}")
 
             logger.info(
-                f"File '{CONVERTED_FILE_NAME}' and '{file.filename}'deleted successfully."
+                f"File '{CONVERTED_FILE_NAME}' and '{safe_filename}' deleted successfully."
             )
         except Exception as e:
             logger.error(e)
